eliminated most GCC 4.3 warnings
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@1411 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ src/cvaux/vs/blobtrackingmsfg.cpp
@@ -788,7 +788,7 @@ public:
         if(m_ParticleNum>0)
         {
             Realloc();
-            printf("sizeof(DefParticle) is %d\n", sizeof(DefParticle));
+            printf("sizeof(DefParticle) is %d\n", (int)sizeof(DefParticle));
             cvReadStructByName(fs,node,"ParticlesPredicted",m_pParticlesPredicted,"ffffiffd");
             cvReadStructByName(fs,node,"ParticlesResampled",m_pParticlesResampled,"ffffiffd");
         }
